refactor redux stuff
rename files:
- reducer.ts to noteReducer.ts;
- actions.ts to NoteActionCreators.ts;

change:
- move RootState type from reducer.ts to store.ts and change the type
  definition;
- update RootState type usage in useTypedSelector.ts;
- update usage of functions imported from actions.ts
  (NoteActionCreators.ts);
