+++ src/hooks/useTypedSelector.ts
@@ -1,5 +1,5 @@
 import { TypedUseSelectorHook, useSelector } from 'react-redux';
-import { RootState } from '../redux/reducer';
+import { RootState } from '../redux/store';
 
 const useTypedSelector: TypedUseSelectorHook<RootState> = useSelector;
